Updating code for Page Builder to be more DRY and allowing fields to be built based on API using a react context.
