+++ frontend/packages/gatsby-plugin-tinacms-pagebuilder/src/pages/edit/new/[...].js
@@ -1,26 +1,20 @@
-import React, {useEffect} from "react"
-import { graphql, navigate, useStaticQuery } from "gatsby"
+import React from "react"
+import { graphql, useStaticQuery } from "gatsby"
 import set from "lodash.set"
-import { imageListBlock } from "gatsby-theme-core-design-system/src/components/Images/ImageList"
-import { paragraphBlock } from "gatsby-theme-core-design-system/src/components/Text/Paragraph"
-import { featureListBlock } from "gatsby-theme-core-design-system/src/components/Features/FeatureList"
-import { accordionListBlock } from "gatsby-theme-core-design-system/src/components/Accordions/AccordionList"
-import { signpostListBlock } from "gatsby-theme-core-design-system/src/components/Signposts/SignpostList"
-import { sliderListBlock } from "gatsby-theme-core-design-system/src/components/Sliders/SliderList"
-import {Hero, heroBlock} from "gatsby-theme-core-design-system/src/components/Heros/Hero"
-import {BlocksControls, InlineBlocks, InlineForm, InlineWysiwyg} from "react-tinacms-inline"
+import { InlineBlocks, InlineForm } from "react-tinacms-inline"
 import { InitForm, InitPlugin } from "../../../../utils/inits";
-import { isTinaWindow, createTinaField } from "../../../../utils/tinaUtils";
+import {
+  createBlockComponent,
+  createTinaField,
+  createTinaInlineBlocks,
+  isTinaWindow
+} from "../../../../utils/tinaUtils";
 import Seo from "gatsby-theme-core-design-system/src/components/Misc/Seo"
-import { isLoggedIn } from "@powerstack/drupal-oauth-connector"
 import Header from "gatsby-theme-core-design-system/src/components/Headers/Header"
 import Footer from "gatsby-theme-core-design-system/src/components/Footers/Footer"
 import { Title } from "gatsby-theme-core-design-system/src/components/Text/Title"
-import { Box, Spinner, Text } from "theme-ui"
-import {submitTinaDataToDrupal} from "../../../../utils/postRequestUtils";
-import { createTinaInlineBlocks, getRequestFetchMultiple} from "../../../../utils/getRequestUtils";
-import {capitalize} from "@powerstack/utils";
-
+import { Box, Spinner } from "theme-ui"
+import { getRequestFetchMultiple } from "../../../../utils/getRequestUtils";
 
 export const PageBuilderContext = React.createContext()
 
@@ -41,62 +35,16 @@ const NewPage = ({ serverData }) => {
       }
     }
   `)
+
   console.log(serverData)
-  const createComponent = ({ index, data }) => {
-    if(React.useContext(PageBuilderContext)[data.path] && React.useContext(PageBuilderContext)[data.path].hasOwnProperty("children")) {
-      const blockKey = Object.keys(React.useContext(PageBuilderContext)[data.path].children)[0]
-      const availableBlocks = {
-      }
-      set(availableBlocks, `${blockKey}`, Object.values(React.useContext(PageBuilderContext)[data.path].children)[0])
-      set(availableBlocks, `${blockKey}.Component`, createComponent)
 
-      console.log(availableBlocks)
-
-      return(
-        <BlocksControls index={index} focusRing={{ offset: 0 }} insetControls>
-          <Section />
-          <InlineBlocks className={`-blocks`} name={`blocks`} blocks={availableBlocks} />
-        </BlocksControls>
-      )
-    } else {
-      return(
-        <BlocksControls index={index} focusRing={{ offset: 0 }} insetControls>
-          <InlineWysiwyg name="text" format="html">
-            <p
-              className="paragraph__text"
-              dangerouslySetInnerHTML={{
-                __html: data.text,
-              }}
-            />
-            <p>Text</p>
-          </InlineWysiwyg>
-
-        </BlocksControls>
-      )
-    }
-    console.log(data)
-    console.log()
-
-  }
-  const objKey=Object.keys(serverData.blocks)[0]
-  set(serverData, `blocks.${objKey}.Component`, createComponent)
+  set(serverData, `blocksData.${Object.keys(serverData.blocksData)[0]}.Component`, createBlockComponent)
 
   const PageForm = {
     initialValues: {
       status: true
     },
     fields: serverData.fields
-  }
-
-  const [, pageForm] = isTinaWindow ? InitForm(PageForm) : ["", ""]
-
-  isTinaWindow && InitPlugin(pageForm)
-
-  const formConfig = {
-    initialValues: serverData.content,
-    onSubmit(data) {
-      submitTinaDataToDrupal(data)
-    },
   }
 
   const [, form] = isTinaWindow ? InitForm(PageForm) : ["", ""]
@@ -119,10 +67,10 @@ const NewPage = ({ serverData }) => {
       <Header siteTitle={data.site.siteMetadata?.title || `Title`} />
       <div className="home">
         {isTinaWindow ? (
-          <PageBuilderContext.Provider value={serverData.blocks}>
+          <PageBuilderContext.Provider value={serverData.blocksData}>
             <InlineForm form={isTinaWindow && form}>
               <Title title={serverData.content?.title} />
-              <InlineBlocks className={'blocks'} name="blocks" blocks={serverData.blocks} />
+              <InlineBlocks className={'blocks'} name="blocks" blocks={serverData.blocksData} />
             </InlineForm>
           </PageBuilderContext.Provider>
         ) : (
@@ -144,13 +92,6 @@ const NewPage = ({ serverData }) => {
   )
 }
 
-export function Section ({ index, data }) {
-  return (
-    <div>test</div>
-  )
-}
-
-
 /**
  * Gatsby API's get server data function
  * @param params
@@ -160,6 +101,7 @@ export function Section ({ index, data }) {
 export async function getServerData({ params, headers }) {
   const [type, templateId] = params['*'].split('/')
 
+  // An object of the requests we are going to execute against Drupal's API
   const requests = {
     contentTypeData: `entity_form_display/entity_form_display?filter[bundle]=${type}`,
     paragraphsData: 'entity_form_display/entity_form_display?filter[targetEntityType]=paragraph',
@@ -168,12 +110,16 @@ export async function getServerData({ params, headers }) {
     apiBaseData: ''
   }
 
+  // Execute the requests
   const requestsData = await getRequestFetchMultiple(headers, requests)
 
-  const pageFields = []
-  let pageBlocks = {}
-
+  // Make sure that we don't have any errors returned
   if(requestsData.errors && Object.keys(requestsData.errors).length === 0 && Object.getPrototypeOf(requestsData.errors) === Object.prototype) {
+    // Local variables
+    const pageFields = []
+    let blocksData = {}
+
+    // Set the current Drupal UID
     const currentUserUuid = requestsData.success.apiBaseData.meta.links.me.meta.id
     let currentUserId = 0
 
@@ -184,6 +130,7 @@ export async function getServerData({ params, headers }) {
       }
     })
 
+    // Iterate over data to build out form fields and
     for (const entry of Object.entries(requestsData.success.contentTypeData.data[0].attributes.content).sort((a, b) => (a[1].weight > b[1].weight) ? 1 : -1 )){
       const [key, value] = entry;
       if(!value.type.startsWith("entity_reference_paragraphs")){
@@ -193,8 +140,7 @@ export async function getServerData({ params, headers }) {
         }
       } else {
         const fieldData = await createTinaInlineBlocks(type, value, headers)
-
-        pageBlocks = fieldData
+        blocksData = fieldData
       }
     }
 
@@ -202,8 +148,8 @@ export async function getServerData({ params, headers }) {
       props: {
         ...requestsData.success,
         currentUser: currentUserId,
-        blocks: pageBlocks,
-        fields: pageFields
+        fields: pageFields,
+        blocksData
       },
     }
   } else {
